+++ src/manifest.class.ts
@@ -4,6 +4,7 @@ import { IM3u8Line } from './interfaces/m3u8-line.interface'
 import { IM3u8Producer } from './interfaces/m3u8-producer.interface'
 import { IM3u8Variant } from './interfaces/m3u8-variant.interface'
 
+export type VariantResolver = (source: string) => string
 export type VariantModifier = (variant: IM3u8Variant) => IM3u8Variant
 export type VariantPredicate = (variant: IM3u8Variant) => boolean