+++ test/media-m3u8.spec.ts
@@ -1,6 +1,7 @@
 
 import { M3u8LineType } from '../src/enums/m3u8-line-type.enum'
 import { MediaM3u8 } from '../src/media-m3u8.class'
+import { TargetResolver } from '../src/util/target-resolver.util'
 
 const SPECIMEN = `
 #EXTM3U
@@ -31,7 +32,38 @@ const SPECIMEN = `
 #EXTINF:2.068733,
 960x540-2000kbps_211f1e36f293d1b6.ts
 #EXT-X-ENDLIST
-`
+`.trim()
+
+const RESOLVED_SPECIMEN = `
+#EXTM3U
+#EXT-X-VERSION:6
+#EXT-X-TARGETDURATION:6
+#EXT-X-MEDIA-SEQUENCE:0
+#EXT-X-PLAYLIST-TYPE:VOD
+#EXT-X-INDEPENDENT-SEGMENTS
+#EXT-X-KEY:METHOD=AES-128,URI="/keys/12345/encryption.key",IV=0xb7cb82dd5e12261c81eb13eba84e9ca3
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_f8a1ae2a59a2b2f7.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_23222e7e8b6b1494.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_457d5113a92ea5b6.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_f2d29dd65a052b14.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_8b54a11735c40542.ts
+#EXTINF:5.972633,
+https://example.com/12345/960x540-2000kbps_7461ea56e5f5d964.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_5445f203bad03f6c.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_fd33e3f4aed0ae41.ts
+#EXTINF:6.006000,
+https://example.com/12345/960x540-2000kbps_431a5b4097d066e0.ts
+#EXTINF:2.068733,
+https://example.com/12345/960x540-2000kbps_211f1e36f293d1b6.ts
+#EXT-X-ENDLIST
+`.trim()
 
 describe('parsing a media m3u8 file', () => {
 
@@ -71,5 +103,14 @@ describe('parsing a media m3u8 file', () => {
   it('should correctly identify the endlist segment', () => {
     expect(example.lines[example.lines.length - 1].type).toBe(M3u8LineType.FOOTER)
     expect(example.lines[example.lines.length - 1].content).toBe('#EXT-X-ENDLIST')
+  })
+
+  it('should correctly resolve the media m3u8', () => {
+    const resolver = new TargetResolver(
+      (value) => value.replace('encryption.key', '/keys/12345/encryption.key'),
+      (value) => `https://example.com/12345/${value}`
+    )
+    const resolved = example.resolve(resolver).marshal()
+    expect(resolved).toBe(RESOLVED_SPECIMEN)
   })
 })